tests/wpt/web-platform-tests/workers/WorkerGlobalScope_ErrorEvent_colno.htm

<!DOCTYPE html>
<title> WorkerGlobalScope onerror event handler argument: col </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var worker = new Worker('./support/ErrorEvent.js');
  worker.onmessage = t.step_func_done(function(e) {
    assert_equals(typeof e.data.colno, "number");
  });
  worker.postMessage("Error Message");
});
</script>
